Move some Workspace contracts into LanguageService
