+++ src/Microsoft.SqlTools.LanguageService/Localization/sr.cs
@@ -31,6 +31,14 @@ namespace Microsoft.SqlTools.LanguageService
         }
 
 
+        public static string WorkspaceServicePositionLineOutOfRange
+        {
+            get
+            {
+                return Keys.GetString(Keys.WorkspaceServicePositionLineOutOfRange);
+            }
+        }
+
         public static string ErrorEmptyStringReplacement
         {
             get
@@ -44,6 +52,11 @@ namespace Microsoft.SqlTools.LanguageService
             return Keys.GetString(Keys.WorkspaceServiceBufferPositionOutOfOrder, sLine, sCol, eLine, eCol);
         }
 
+        public static string WorkspaceServicePositionColumnOutOfRange(int line)
+        {
+            return Keys.GetString(Keys.WorkspaceServicePositionColumnOutOfRange, line);
+        }
+
         [System.Runtime.CompilerServices.CompilerGeneratedAttribute()]
         public class Keys
         {
@@ -53,6 +66,12 @@ namespace Microsoft.SqlTools.LanguageService
 
 
             public const string WorkspaceServiceBufferPositionOutOfOrder = "WorkspaceServiceBufferPositionOutOfOrder";
+
+
+            public const string WorkspaceServicePositionLineOutOfRange = "WorkspaceServicePositionLineOutOfRange";
+
+
+            public const string WorkspaceServicePositionColumnOutOfRange = "WorkspaceServicePositionColumnOutOfRange";
 
 
             public const string ErrorEmptyStringReplacement = "ErrorEmptyStringReplacement";
@@ -76,6 +95,12 @@ namespace Microsoft.SqlTools.LanguageService
             public static string GetString(string key)
             {
                 return resourceManager.GetString(key, _culture);
+            }
+
+
+            public static string GetString(string key, object arg0)
+            {
+                return string.Format(global::System.Globalization.CultureInfo.CurrentCulture, resourceManager.GetString(key, _culture), arg0);
             }
