+++ src/Microsoft.SqlTools.LanguageService/Localization/sr.resx
@@ -122,6 +122,15 @@
     <comment>.
  Parameters: 0 - sLine (int), 1 - sCol (int), 2 - eLine (int), 3 - eCol (int) </comment>
   </data>
+  <data name="WorkspaceServicePositionLineOutOfRange" xml:space="preserve">
+    <value>Position is outside of file line range</value>
+    <comment></comment>
+  </data>
+  <data name="WorkspaceServicePositionColumnOutOfRange" xml:space="preserve">
+    <value>Position is outside of column range for line {0}</value>
+    <comment>.
+ Parameters: 0 - line (int) </comment>
+  </data>
   <data name="ErrorEmptyStringReplacement" xml:space="preserve">
     <value>Replacement of an empty string by an empty string.</value>
     <comment></comment>
