+++ src/Microsoft.SqlTools.LanguageService/Microsoft.SqlTools.LanguageService.csproj
@@ -33,5 +33,8 @@
     <InternalsVisibleTo Include="Microsoft.SqlTools.LanguageService.UnitTests" />
     <InternalsVisibleTo Include="MicrosoftSqlToolsServiceLayer" />
     <InternalsVisibleTo Include="Microsoft.SqlTools.ServiceLayer.UnitTests" />
+    <InternalsVisibleTo Include="Microsoft.SqlTools.ServiceLayer.IntegrationTests" />
+    <InternalsVisibleTo Include="Microsoft.SqlTools.ServiceLayer.Test.Common" />
+    <InternalsVisibleTo Include="DynamicProxyGenAssembly2" />
   </ItemGroup>
 </Project>
