+++ src/Microsoft.SqlTools.LanguageService/Workspace/Workspace.cs
@@ -11,11 +11,10 @@ using System.Text;
 using System.Text.RegularExpressions;
 using System.Linq;
 using Microsoft.SqlTools.Utility;
-using Microsoft.SqlTools.ServiceLayer.Workspace.Contracts;
-using Microsoft.SqlTools.ServiceLayer.Utility;
+using Microsoft.SqlTools.LanguageService.Workspace.Contracts;
 using System.Collections.Concurrent;
 
-namespace Microsoft.SqlTools.ServiceLayer.Workspace
+namespace Microsoft.SqlTools.LanguageService.Workspace
 {
     /// <summary>
     /// Manages a "workspace" of script files that are open for a particular
@@ -137,7 +136,7 @@ namespace Microsoft.SqlTools.ServiceLayer.Workspace
                 }
 
                 // Get the absolute file path
-                ResolvedFile resolvedFile = FileUtilities.TryGetFullPath(filePath, clientUri);
+                ResolvedFile resolvedFile = TryGetFullPath(filePath, clientUri);
                 filePath = resolvedFile.FilePath;
                 canReadFromDisk = resolvedFile.CanReadFromDisk;
             }
@@ -145,6 +144,26 @@ namespace Microsoft.SqlTools.ServiceLayer.Workspace
             Logger.Verbose("Resolved path: " + clientUri);
 
             return new ResolvedFile(filePath, clientUri, canReadFromDisk);
+        }
+
+        /// <summary>
+        /// Attempts to resolve the given filePath to an absolute path to a file on disk,
+        /// defaulting to the original filePath if that fails.
+        /// </summary>
+        /// <param name="filePath">The file path to resolve</param>
+        /// <param name="clientUri">The full file path URI used by the client</param>
+        /// <returns></returns>
+        private static ResolvedFile TryGetFullPath(string filePath, string clientUri)
+        {
+            try
+            {
+                return new ResolvedFile(Path.GetFullPath(filePath), clientUri, true);
+            }
+            catch (NotSupportedException)
+            {
+                // This is not a standard path.
+                return new ResolvedFile(filePath, clientUri, false);
+            }
         }
 
          /// <summary>
