+++ src/Microsoft.SqlTools.ServiceLayer/Localization/sr.strings
@@ -230,13 +230,6 @@ StarExpansionLabel(string starExpression) = Expand {0}
 StarExpansionDescription(string starExpression, string columnCount, string columnPreview) = Replace {0} with {1} columns: {2}
 
 ############################################################################
-# Workspace Service
-
-WorkspaceServicePositionLineOutOfRange = Position is outside of file line range
-
-WorkspaceServicePositionColumnOutOfRange(int line) = Position is outside of column range for line {0}
-
-############################################################################
 # Edit Data Service
 
 EditDataObjectNotFound = Table or view requested for edit could not be found
